[Tests] move helpers in to `test/_helpers` dir
